+++ ImgPlane.h
@@ -31,11 +31,11 @@ struct ImgPlane {
     Halide::Runtime::Buffer<double, 2> pixel_locs;
 };
 
-const int N_HAT[] = { 0, 0, 1 };
+const float N_HAT[] = { 0, 0, 1 };
 
 ImgPlane img_plane_create(PlatformData &pd,
                           double res_factor = 1.0,
-                          const int n_hat[3] = N_HAT,
+                          const float n_hat[3] = N_HAT,
                           double aspect = 1.0,
                           bool upsample = true);
